+++ tests/test_perovskite.py
@@ -49,8 +49,37 @@ def test_solved_values():
     os.remove(result_file)
 
 
+def test_solved_values_experimental():
+    # setup
+    solve_interaction(
+        structure=structure_file,
+        orbitals_dict=orbitals,
+        rcut=rcut,
+        save_filename=result_file,
+        experimental=True
+    )
+    # action
+    relationship = OrbitalPropertyRelationship.from_file(result_file)
+    retrived_values = find_interaction_values(
+        stored_interaction,
+        [relationship.all_pairs[i] for i in relationship.free_pair_indices]
+    )
+    tb = relationship.get_tightbinding_from_free_parameters(
+        free_Hijs=retrived_values, experimental=True)
+    #unitcell = kpt.UnitCell(model.tb.cell)
+    #kpath = unitcell.get_kpath_from_path_string(bandpath)
+    kpath = reciprocal.Kpath.from_cell_pathstring(tb.cell, bandpath)
+    values, _ = tb.solveE_at_ks(kpath.kpoints)
+    stacked = np.hstack([kpath.kpoints, values])
+    # test
+    compare = np.load(stored_result)
+    assert np.allclose(stacked, compare, atol = 1e-3)
+    os.remove(result_file)
+
+
 if __name__ == "__main__":
     from automaticTB.properties import bands
+    from automaticTB import tools
     solve_interaction(
         structure=structure_file,
         orbitals_dict=orbitals,
@@ -62,6 +91,8 @@ if __name__ == "__main__":
         stored_interaction,
         [relationship.all_pairs[i] for i in relationship.free_pair_indices]
     )
+    relationship.print_free_pairs()
+    print(retrived_values)
     tb = relationship.get_tightbinding_from_free_parameters(
         free_Hijs=retrived_values)