+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Since the project is still in development, changes are logged without version number
 
+## 2023-04-09
+
+### Changed
+
+- In `analysis/calculation` array nk is changed to list nk.
+
 ## 2023-04-07
 
 Major change to the properties subpackage.
@@ -31,3 +37,10 @@ Major change to the properties subpackage.
 - Class object defined in `transport.py` is removed. use `calculate_transport()` instead
 - `UnitCell` in kpoints.py (reciprocal.py) is now removed
 - `ElectronicModel` class is now removed
+
+## 2023-04-30
+
+### Added
+
+- added an experimental solver that should treat the conjugate relationship correctly.
+- `LinearEquation` has a method now to return the solvable part of the homogeneous matrix.